add description of m and n, fix typos (#415)
Add to previous changes for PR closing issue #329.
Previous changes added description of the range of the FriedmanRafsky
test in terms of `m` and `n` but neglected to mention what `m` and `n`
are. This commit adds a brief description of `m` and `n`, as well as
correcting some typos that were noticed along the way.